dpkg_lowpkg: Do not access /var/lib/dpkg/info/<package>.list directly
salt contains modules, which directly access the dpkg internal database,
instead of using one of the public interfaces provided by dpkg. This is
a problem for several reasons, because even though the layout and format
of the dpkg database is administrator friendly, and it is expected that
those might need to mess with it, in case of emergency, this “interface”
does not extend to other programs besides the dpkg suite of tools. The
admindir can also be configured differently at dpkg build or run-time.
And finally, the contents and its format, will be changing in the near
future.

So use something like:

```
dpkg-query --showformat '${db-fsys:Last-Modified}\n' --show $pkg
```

to get the mtime from /var/lib/dpkg/info/<package>.list files.

Bug: https://github.com/saltstack/salt/issues/52605
Bug-Debian: https://bugs.debian.org/944970
Forwarded: https://github.com/saltstack/salt/pull/58735
Signed-off-by: Benjamin Drung <benjamin.drung@ionos.com>
Gbp-Pq: Name dpkg_lowpkg-Do-not-access-var-lib-dpkg-info-package-.patch