linux_sysctl.present: Fix replacing spaces by tabs
`salt/modules/linux_sysctl.py` tries to be smart in `persist()` and
replaces spaces by tabs, which will be correct for keys like
`net.ipv4.tcp_rmem` but not for `kernel.core_pattern`.

So only sanitize the values for comparison, but not for persisting them
in the configuration file.

Bug: https://github.com/saltstack/salt/issues/40054
Forwarded: https://github.com/saltstack/salt/pull/61452
Signed-off-by: Benjamin Drung <benjamin.drung@ionos.com>